ml-service/requirements.txt
flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
google-genai==1.0.0
elevenlabs==1.59.0
